Move job cache framework into lib-orch
