+++ tracdap-libs/tracdap-lib-orch/src/main/java/org/finos/tracdap/common/cache/local/LocalJobCache.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Accenture Global Solutions Limited
+ * Copyright 2023 Accenture Global Solutions Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.svc.orch.cache.local;
+package org.finos.tracdap.common.cache.local;
 
+import org.finos.tracdap.common.cache.CacheQueryResult;
+import org.finos.tracdap.common.cache.IJobCache;
+import org.finos.tracdap.common.cache.Ticket;
 import org.finos.tracdap.common.exception.ECacheNotFound;
 import org.finos.tracdap.common.exception.ECacheTicket;
-import org.finos.tracdap.svc.orch.cache.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
