+++ tracdap-libs/tracdap-lib-orch/src/main/java/org/finos/tracdap/common/cache/local/LocalJobCacheEntry.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Accenture Global Solutions Limited
+ * Copyright 2023 Accenture Global Solutions Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.svc.orch.cache.local;
+package org.finos.tracdap.common.cache.local;
 
-import org.finos.tracdap.svc.orch.cache.Ticket;
+import org.finos.tracdap.common.cache.Ticket;
 
 import java.time.Instant;
